Rename featurestore* -> feature_store*.
