+++ searchlib/src/vespa/searchlib/memoryindex/feature_store.cpp
@@ -1,6 +1,6 @@
 // Copyright 2017 Yahoo Holdings. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
 
-#include "featurestore.h"
+#include "feature_store.h"
 #include <vespa/searchlib/index/schemautil.h>
 #include <vespa/searchlib/datastore/datastore.hpp>
